refactor: findArticlesController 인터페이스와 여러 구현체들로 분리
