+++ backend/footprint/src/main/java/foot/footprint/domain/article/application/findArticles/FindArticlesService.java
@@ -1,10 +1,10 @@
-package foot.footprint.domain.article.application.findArticle;
+package foot.footprint.domain.article.application.findArticles;
 
 import foot.footprint.domain.article.domain.LocationRange;
 import foot.footprint.domain.article.dto.ArticleMapResponse;
 import java.util.List;
 
-public interface FindArticleService {
+public interface FindArticlesService {
 
     public List<ArticleMapResponse> findArticles(Long memberId, Long groupId,
         LocationRange locationRange);
